Reorganize tests.
